theoretically a wall follower

+++ 2020-2021/botCode/sensorsActuators/sa_interface/tofsensor.hpp
@@ -1,20 +1,21 @@
 #ifndef TOFSENSOR_H
 #define TOFSENSOR_H
 
-#include <vl6180_pi/vl6180_pi.h>
 #include <string>
+#include <vl6180_pi/vl6180_pi.h>
 
 using namespace std;
 
 /* Time of Flight Sensors - Specifically vl6180_x tof sensors*/
-class TofSensor{
-	public:
-		TofSensor(int pin, string name, int address);
-		int distance();
-		int pin;
-		string name;
-		int address;
-	private:
-		vl6180 handle;
+class TofSensor {
+  public:
+    TofSensor(int pin, string name, int address);
+    double distance();
+    int pin;
+    string name;
+    int address;
+
+  private:
+    vl6180 handle;
 };
 #endif
